fix: Repair the create-organization modal test, which still pointed at the pre-rename path and asserted the validation toast that the disabled submit button replaced. It now mirrors the create-project test.
